+++ MatKollen/ViewModels/GroceryListViewModel.cs
@@ -10,7 +10,7 @@ namespace MatKollen.ViewModels
     {
         public required ListFoodItem FoodDetails  { get; set; }
 
-        public double ConvertedQuantity { get; set; }
+        public decimal ConvertedQuantity { get; set; }
         public required string FoodItemName { get; set; }
         public required string Unit { get; set; }
         public required string ListName { get; set; }
